payatom_bot/requirements.txt
python-telegram-bot==20.7
selenium==4.23.1
requests==2.32.3
python-dotenv==1.0.1
